Adding desc to "No" state and forced wait time to loading
I decided to add some text to explain what "No" meant, since I don't think it's very clear what this site does without it.  The "Yes/Maybe" states will have the weather alert data on screen, so that will be obvious.
Also adding some "load-time" to the app so that we can see the clever messages.  Why go through all the trouble of writing clever loading messages if the API is too fast to see them (and Andrew was pretty sad)